# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectedResponse": {
        "properties": {
          "connected": {
            "title": "Connected",
            "type": "boolean"
          }
        },
        "required": [
          "connected"
        ],
        "title": "ConnectedResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HoldingResponse": {
        "properties": {
          "absolute_excess_pp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Absolute Excess Pp"
          },
          "absolute_inr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Absolute Inr"
          },
          "absolute_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Absolute Pct"
          },
          "avg_price": {
            "title": "Avg Price",
            "type": "number"
          },
          "benchmark": {
            "title": "Benchmark",
            "type": "string"
          },
          "benchmark_return": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Benchmark Return"
          },
          "cagr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cagr"
          },
          "cagr_excess_pp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cagr Excess Pp"
          },
          "incomplete": {
            "title": "Incomplete",
            "type": "boolean"
          },
          "instrument_id": {
            "title": "Instrument Id",
            "type": "integer"
          },
          "instrument_type": {
            "title": "Instrument Type",
            "type": "string"
          },
          "ltp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ltp"
          },
          "mf_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mf Category"
          },
          "needs_category": {
            "title": "Needs Category",
            "type": "boolean"
          },
          "qty": {
            "title": "Qty",
            "type": "number"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "windows": {
            "$ref": "#/components/schemas/WindowsMap"
          },
          "xirr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Xirr"
          },
          "xirr_excess_pp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Xirr Excess Pp"
          }
        },
        "required": [
          "instrument_id",
          "symbol",
          "instrument_type",
          "qty",
          "avg_price",
          "ltp",
          "value",
          "absolute_pct",
          "absolute_inr",
          "xirr",
          "cagr",
          "benchmark",
          "benchmark_return",
          "absolute_excess_pp",
          "xirr_excess_pp",
          "cagr_excess_pp",
          "incomplete",
          "needs_category",
          "mf_category",
          "windows"
        ],
        "title": "HoldingResponse",
        "type": "object"
      },
      "HoldingSeriesResponse": {
        "additionalProperties": false,
        "properties": {
          "as_of": {
            "title": "As Of",
            "type": "string"
          },
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "benchmark": {
            "title": "Benchmark",
            "type": "string"
          },
          "incomplete": {
            "title": "Incomplete",
            "type": "boolean"
          },
          "instrument_id": {
            "title": "Instrument Id",
            "type": "integer"
          },
          "metric": {
            "const": "absolute",
            "title": "Metric",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/SeriesPoint"
            },
            "title": "Points",
            "type": "array"
          },
          "start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "window": {
            "enum": [
              "ITD",
              "1Y",
              "3Y",
              "5Y"
            ],
            "title": "Window",
            "type": "string"
          }
        },
        "required": [
          "instrument_id",
          "window",
          "metric",
          "benchmark",
          "as_of",
          "start",
          "available",
          "incomplete",
          "points"
        ],
        "title": "HoldingSeriesResponse",
        "type": "object"
      },
      "HoldingTransactionsResponse": {
        "additionalProperties": false,
        "properties": {
          "instrument_id": {
            "title": "Instrument Id",
            "type": "integer"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionRowResponse"
            },
            "title": "Transactions",
            "type": "array"
          }
        },
        "required": [
          "instrument_id",
          "transactions"
        ],
        "title": "HoldingTransactionsResponse",
        "type": "object"
      },
...
      "PortfolioSeriesResponse": {
        "additionalProperties": false,
        "properties": {
          "as_of": {
            "title": "As Of",
            "type": "string"
          },
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "incomplete": {
            "title": "Incomplete",
            "type": "boolean"
          },
          "metric": {
            "const": "absolute",
            "title": "Metric",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/SeriesPoint"
            },
            "title": "Points",
            "type": "array"
          },
          "start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "window": {
            "enum": [
              "ITD",
              "1Y",
              "3Y",
              "5Y"
            ],
            "title": "Window",
            "type": "string"
          }
        },
        "required": [
          "window",
          "metric",
          "as_of",
          "start",
          "available",
          "incomplete",
          "points"
        ],
        "title": "PortfolioSeriesResponse",
        "type": "object"
      },
...
      "RequestTokenBody": {
        "properties": {
          "request_token": {
            "title": "Request Token",
            "type": "string"
          }
        },
        "required": [
          "request_token"
        ],
        "title": "RequestTokenBody",
        "type": "object"
      },
      "SeriesPoint": {
        "additionalProperties": false,
        "properties": {
          "benchmark_return": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Benchmark Return"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "holding_return": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Holding Return"
          },
          "portfolio_return": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portfolio Return"
          }
        },
        "required": [
          "date"
        ],
        "title": "SeriesPoint",
        "type": "object"
      },
...
      "TransactionRowResponse": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "fees": {
            "title": "Fees",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "price": {
            "title": "Price",
            "type": "number"
          },
          "quantity": {
            "title": "Quantity",
            "type": "number"
          },
          "side": {
            "title": "Side",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "trade_date": {
            "title": "Trade Date",
            "type": "string"
          }
        },
        "required": [
          "id",
          "trade_date",
          "side",
          "quantity",
          "price",
          "fees",
          "amount",
          "source"
        ],
        "title": "TransactionRowResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WindowMetrics": {
        "additionalProperties": false,
        "properties": {
          "absolute_excess_pp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Absolute Excess Pp"
          },
          "absolute_inr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Absolute Inr"
          },
          "absolute_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Absolute Pct"
          },
          "benchmark_return": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Benchmark Return"
          },
          "cagr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cagr"
          },
          "cagr_excess_pp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cagr Excess Pp"
          },
          "xirr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Xirr"
          },
          "xirr_excess_pp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Xirr Excess Pp"
          }
        },
        "title": "WindowMetrics",
        "type": "object"
      },
      "WindowsMap": {
        "additionalProperties": false,
        "description": "Fixed window keys so OpenAPI/TS keep named properties (not string index).",
        "properties": {
          "1Y": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WindowMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "3Y": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WindowMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "5Y": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WindowMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "ITD": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WindowMetrics"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "ITD",
          "1Y",
          "3Y",
          "5Y"
        ],
        "title": "WindowsMap",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Portfolio Tracker API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/auth/callback": {
      "get": {
        "operationId": "callback_get_auth_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "request_token",
            "required": true,
            "schema": {
              "title": "Request Token",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Callback Get",
        "tags": [
          "auth"
        ]
      },
      "post": {
        "operationId": "callback_auth_callback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestTokenBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Callback",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/portfolio/holdings/{instrument_id}": {
      "get": {
        "operationId": "holding_detail_portfolio_holdings__instrument_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "instrument_id",
            "required": true,
            "schema": {
              "title": "Instrument Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Holding not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Holding Detail",
        "tags": [
          "portfolio"
        ]
      }
    },
    "/portfolio/holdings/{instrument_id}/series": {
      "get": {
        "operationId": "holding_series_portfolio_holdings__instrument_id__series_get",
        "parameters": [
          {
            "in": "path",
            "name": "instrument_id",
            "required": true,
            "schema": {
              "title": "Instrument Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": "ITD",
              "enum": [
                "ITD",
                "1Y",
                "3Y",
                "5Y"
              ],
              "title": "Window",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldingSeriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid window"
          },
          "404": {
            "description": "Instrument not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Holding Series",
        "tags": [
          "portfolio"
        ]
      }
    },
    "/portfolio/holdings/{instrument_id}/transactions": {
      "get": {
        "operationId": "holding_transactions_portfolio_holdings__instrument_id__transactions_get",
        "parameters": [
          {
            "in": "path",
            "name": "instrument_id",
            "required": true,
            "schema": {
              "title": "Instrument Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldingTransactionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Instrument not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Holding Transactions",
        "tags": [
          "portfolio"
        ]
      }
    },
...
    "/portfolio/series": {
      "get": {
        "operationId": "portfolio_series_portfolio_series_get",
        "parameters": [
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": "ITD",
              "enum": [
                "ITD",
                "1Y",
                "3Y",
                "5Y"
              ],
              "title": "Window",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioSeriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid window"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Portfolio Series",
        "tags": [
          "portfolio"
        ]
      }
    },
...
  }
}